Component.vue, License.vue, Product.vue, Project.vue all show all of the data.
Fixed the toDateString() problem that created the long version of a date. Date is now in form of 2017-11-12 for example. Sleeker, and contains all that is needed.

Fixed a small bug, that made the page where showing 1 license, not show all of the correct info.

Search by name implemented for all xList pages. Now it only has to be implemented in the "Add X" pages.